fix: fix finding libpython from a venv in msys2
This adds a hack to CMaker to find libpython when it is called inside
of a virtualenv runnin in a MSYS envrionment on Windows.

The `LIBRARY` sysconfig variable returns `libpython3.10.a` in this case
and the variables like `LIBDIR` point to the virtualenv which doesn't
contain the library. Furthermore, the actual file on disk
`libpython3.10.dll.a` which doesn't match any of the existing
combinations.

The virtualenv doesn't override the `prefix` variable from `sysconfig`
so we can use this to find the actual location on disk.